Use react for settings popup
