+++ package.json
@@ -18,7 +18,13 @@
     "url": "https://github.com/material-extensions/material-icons-browser-extension/issues"
   },
   "dependencies": {
+    "@emotion/react": "11.13.3",
+    "@emotion/styled": "11.13.0",
+    "@mui/icons-material": "6.1.2",
+    "@mui/material": "6.1.2",
     "material-icon-theme": "5.11.1",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "selector-observer": "2.1.6",
     "webextension-polyfill": "0.12.0"
   },
@@ -28,9 +34,12 @@
     "@types/fs-extra": "11.0.4",
     "@types/json-stable-stringify": "1.0.36",
     "@types/node": "20.14.10",
+    "@types/react": "18.3.11",
+    "@types/react-dom": "18.3.0",
     "@types/webextension-polyfill": "0.10.7",
     "dotenv": "16.4.5",
     "esbuild": "0.21.5",
+    "esbuild-sass-plugin": "3.3.1",
     "fs-extra": "11.2.0",
     "husky": "9.0.11",
     "json-stable-stringify": "1.1.1",
@@ -44,10 +53,10 @@
   },
   "scripts": {
     "prebuild": "rimraf --glob *.zip ./dist",
-    "build": "run-s build-languages build-src compile bundle",
+    "build": "run-s build-src build-languages check-type-safety bundle",
     "build-languages": "ts-node ./scripts/build-languages.ts",
     "build-src": "ts-node ./scripts/build-src.ts",
-    "compile": "tsc -p ./",
+    "check-type-safety": "tsc -p ./",
     "rebuild-logos": "ts-node ./scripts/build-icons.ts",
     "bundle": "run-p bundle-edge bundle-chrome bundle-firefox",
     "bundle-edge": "zip -r -j github-material-icons-edge-extension.zip dist/chrome-edge",
