+++ src/models/provider.ts
@@ -1,4 +1,4 @@
-export interface Provider {
+export type Provider = {
   name: string;
   domains: { host: string; test: RegExp }[];
   selectors: {
@@ -20,4 +20,6 @@ export interface Provider {
     iconEl: HTMLElement,
     fileName: string
   ) => string;
-}
+};
+
+export type ProviderMap = Record<string, string>;
