+++ src/ui/popup/settings-popup.css
@@ -1,194 +1,8 @@
 body {
   padding: 0;
   margin: 0;
-  color: #1a202c;
-  font-size: 16px;
-  font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-  /** firefox has bigger popup width limit, this is to bring it to parity with chromium **/
-  max-width: 300px;
 }
 
-#content {
-  min-width: 250px;
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  padding: 1.5rem;
-}
-
-#not-supported {
-  display: none;
-}
-
-#domain-settings {
-  display: none;
-}
-
-#settings-header {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-}
-
-#logo {
-  margin-right: auto;
-  height: 24px;
-  width: 24px;
-}
-
-#settings-header h1 {
-  margin: 0;
-  font-size: 1.4rem;
-  font-weight: bold;
-  color: #01579b;
-}
-
-#options-btn {
-  margin-left: auto;
-  width: 24px;
-  height: 24px;
-  filter: grayscale(1);
-}
-
-#options-btn:hover {
-  cursor: pointer;
-  filter: brightness(0.8);
-}
-
-#domain {
-  text-align: center;
-  padding: 0.5rem 0;
-  margin: 0.5rem 0;
-  border-top: solid 1px #dfe1e4;
-  border-bottom: solid 1px #dfe1e4;
-}
-
-#domain h2 {
-  margin: 0;
-  font-size: 1.1rem;
-  font-weight: normal;
-}
-
-#default-disabled-note {
-  color: maroon;
-  margin: 2rem 0;
-  display: none;
-  flex: 1;
-}
-
-#options-link {
-  color: #01579b;
-  cursor: pointer;
-}
-
-#options-link:hover {
-  filter: brightness(1.2);
-}
-
-label {
-  display: block;
-  font-weight: 500;
-  margin: 0 0.75rem 0.5rem 0;
-  line-height: 1.5;
-}
-
-.checkbox-wrapper {
-  display: flex;
-  justify-content: space-between;
-}
-
-.select-wrapper {
-  position: relative;
-  width: 100%;
-  height: fit-content;
-  margin-bottom: 20px;
-}
-
-select {
-  padding: 0 2rem;
-  width: 100%;
-  min-width: 0px;
-  outline: transparent solid 2px;
-  outline-offset: 2px;
-  position: relative;
-  appearance: none;
-  transition: 200ms all;
-  padding-bottom: 1px;
-  line-height: normal;
-  font-size: 1rem;
-  padding-inline-start: 1rem;
-  height: 2.5rem;
-  border-radius: 0.375rem;
-  border-width: 1px;
-  color: #2d3748;
-  border-color: #e2e8f0;
-}
-
-select:hover {
-  border-color: #cbd5e0;
-}
-
-select:focus-visible {
-  border-color: #3182ce;
-  box-shadow: #3182ce 0px 0px 0px 1px;
-}
-
-.select-icon-wrapper {
-  position: absolute;
-  top: 50%;
-  right: 0.5rem;
-  transform: translateY(-50%);
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  width: 1.5rem;
-  height: 100%;
-  color: currentcolor;
-  font-size: 1.25rem;
-  pointer-events: none;
-}
-
-svg.select-icon {
-  width: 1em;
-  height: 1em;
-  color: currentcolor;
-}
-
-#footer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-#provider-wrapper {
-  display: none;
-}
-
-#add-provider {
-  display: none;
-}
-
-#request {
-  display: none;
-}
-
-.btn {
-  color: #5c6068;
-  padding: 0.3rem 1rem;
-  outline: transparent solid 2px;
-  outline-offset: 2px;
-  transition: 200ms all;
-  font-size: 1rem;
-  border-radius: 0.375rem;
-  border-width: 1px;
-  border-color: transparent;
-  background-color: transparent;
-  margin-bottom: 1rem;
-}
-
-.btn:hover {
-  border-color: red;
-  color: red;
-  cursor: pointer;
+.toolbar-spacer {
+  flex: 1 1 auto;
 }
